Created textinputclean and selectinputclean component
